Prune source snapshot tests
