+++ LongevityWorldCup.Tests/AestheticSystemPageTests.cs
@@ -127,215 +127,6 @@ public sealed class AestheticSystemPageTests(TestWebApplicationFactory sharedFac
         }
     }
 
-    [Fact]
-    public async Task AestheticSystem_DefinesSemanticPaletteGeometryAndStateFallbacks()
-    {
-        var factory = sharedFactory;
-        using var client = factory.CreateClient();
-
-        var css = await client.GetStringAsync("/css/aesthetic-system.css");
-
-        Assert.Contains("--lwc-accent: #087685;", css);
-        Assert.Contains("--lwc-on-accent: #ffffff;", css);
-        Assert.Contains("--lwc-on-accent: #082f35;", css);
-        Assert.Contains("--lwc-border-strong: #71808d;", css);
-        Assert.Contains("--lwc-space-4: 1rem;", css);
-        Assert.Contains("--lwc-radius-md: 8px;", css);
-        Assert.Contains("--lwc-shadow-md:", css);
-        Assert.Contains("font-variant-numeric: tabular-nums;", css);
-        Assert.Contains("@media (forced-colors: active)", css);
-        Assert.Contains("@media (prefers-contrast: more)", css);
-        Assert.Contains("@media (prefers-color-scheme: dark)", css);
-        Assert.Contains("@media (prefers-reduced-motion: reduce)", css);
-        Assert.Contains(":read-only", css);
-        Assert.Contains(":not(:placeholder-shown)", css);
-        Assert.Contains(".badge-clickable[title]:focus-visible::after", css);
-        Assert.Contains(".badge-explained[title]:focus-visible::after", css);
-        Assert.Contains("animation-duration: 1ms !important;", css);
-    }
-
-    [Fact]
-    public async Task TaskPages_UsePurposefulVisualsAndReadableCopy()
-    {
-        var factory = sharedFactory;
-        using var client = factory.CreateClient();
-
-        var proofs = await client.GetStringAsync("/proofs");
-        var application = await client.GetStringAsync("/apply");
-        var review = await client.GetStringAsync("/review");
-
-        Assert.Contains("class=\"proof-upload-symbol\"", proofs);
-        Assert.Contains("fa-file-medical", proofs);
-        Assert.DoesNotContain("content-images/proof", proofs);
-        Assert.Contains("id=\"onboardingProofSymbol\" class=\"proof-upload-symbol\"", application);
-        Assert.DoesNotContain("updateIllustration(\"proof\"", application);
-        Assert.Contains("class=\"application-review-visual\"", review);
-        Assert.Contains("bean-waiting.webp?v=", review);
-        Assert.Contains("alt=\"Mr Bean waiting patiently for the review\"", review);
-        Assert.Contains(".proof-upload-copy", proofs);
-        Assert.Contains("text-align: left;", proofs);
-    }
-
-    [Fact]
-    public async Task SharedSystem_UsesOneAccentAndBoundsCompactBadgeDensity()
-    {
-        var factory = sharedFactory;
-        using var client = factory.CreateClient();
-
-        var header = await client.GetStringAsync("/");
-        var badges = await client.GetStringAsync("/js/badges.js");
-
-        Assert.Contains("--secondary-color: var(--primary-color);", header);
-        Assert.Contains("const visibleItems = items.slice(0, 3);", badges);
-        Assert.Contains("const hiddenItems = items.slice(3);", badges);
-    }
-
-    [Fact]
-    public async Task SharedFormAndBadgeStates_KeepAccessibleContrastAndRestrainedMotion()
-    {
-        var factory = sharedFactory;
-        using var client = factory.CreateClient();
-
-        var aestheticCss = await client.GetStringAsync("/css/aesthetic-system.css");
-        var badgeCss = await client.GetStringAsync("/css/badges.css");
-
-        Assert.Contains("color: var(--lwc-muted, #52606d);", aestheticCss);
-        Assert.DoesNotContain("color: #6c7b88;", aestheticCss);
-        Assert.Contains("var(--lwc-duration-fast, 140ms)", badgeCss);
-        Assert.DoesNotContain("shadowPulse", badgeCss);
-        Assert.DoesNotContain("scale(1.2)", badgeCss);
-        Assert.DoesNotContain("animation:shadowPulse", badgeCss);
-    }
-
-    [Fact]
-    public async Task PrivacyPolicy_WrapsExtremeTokensInsideTheSharedVisualSystem()
-    {
-        var factory = sharedFactory;
-        using var client = factory.CreateClient();
-
-        var html = await client.GetStringAsync("/privacy");
-
-        Assert.Contains("/css/aesthetic-system.css?v=", html);
-        Assert.DoesNotContain("<!--AESTHETIC-SYSTEM-->", html);
-        Assert.Contains("p, li {", html);
-        Assert.Contains("overflow-wrap: anywhere;", html);
-        Assert.Contains("word-break: break-word;", html);
-    }
-
-    [Fact]
-    public async Task UnsubscribePage_UsesTheSharedAccentGeometryAndRestrainedMotion()
-    {
-        var factory = sharedFactory;
-        using var client = factory.CreateClient();
-
-        var html = await client.GetStringAsync("/unsubscribe");
-
-        Assert.Contains("background: var(--lwc-accent, #087685);", html);
-        Assert.Contains("color: var(--lwc-on-accent, #ffffff);", html);
-        Assert.Contains("border-radius: var(--lwc-radius-md, 8px);", html);
-        Assert.DoesNotContain("#0a7c0a", html);
-        Assert.DoesNotContain("#096f09", html);
-        Assert.DoesNotContain("data-aos", html);
-        Assert.Contains("color: var(--lwc-success, #1f7a38);", html);
-        Assert.Contains("color: var(--lwc-danger, #b4233b);", html);
-    }
-
-    [Fact]
-    public async Task HelstabChallenge_UsesTheSharedAccentForActionsAndStructuralMarkers()
-    {
-        var factory = sharedFactory;
-        using var client = factory.CreateClient();
-
-        var css = await client.GetStringAsync("/css/helstab-kihivas.css");
-        var html = await client.GetStringAsync("/helstab-kihivas");
-        var mainStart = html.IndexOf("<main class=\"helstab-page\"", StringComparison.Ordinal);
-        Assert.True(mainStart >= 0);
-        var mainEnd = html.IndexOf("</main>", mainStart, StringComparison.Ordinal);
-
-        Assert.True(mainEnd > mainStart);
-        var mainHtml = html[mainStart..(mainEnd + "</main>".Length)];
-
-        Assert.Contains("background: var(--lwc-accent, #087685);", css);
-        Assert.Contains("color: var(--lwc-on-accent, #ffffff);", css);
-        Assert.DoesNotContain("#78da3b", css);
-        Assert.DoesNotContain("#ff4081", css);
-        Assert.DoesNotContain("rgba(120, 218, 59", css);
-        Assert.Contains("fill: currentColor;", css);
-        Assert.DoesNotContain("stroke: currentColor;", css);
-        Assert.Equal(10, mainHtml.Split("class=\"helstab-icon\"", StringSplitOptions.None).Length - 1);
-        Assert.Equal(10, System.Text.RegularExpressions.Regex.Matches(mainHtml, "<svg class=\"helstab-icon\"[^>]*><path ").Count);
-        Assert.Contains("Font Awesome Free 6.7.2", mainHtml);
-        Assert.DoesNotContain("class=\"fas ", mainHtml);
-        Assert.DoesNotContain("class=\"fab ", mainHtml);
-    }
-
-    [Theory]
-    [InlineData("/about")]
-    [InlineData("/history")]
-    [InlineData("/ruleset")]
-    public async Task DocumentationPages_ProgressivelyEnhanceDeepMobileNavigation(string path)
-    {
-        var factory = sharedFactory;
-        using var client = factory.CreateClient();
-
-        var html = await client.GetStringAsync(path);
-
-        Assert.Contains("class=\"documentation-nav-toggle\" aria-expanded=\"false\"", html);
-        Assert.Contains("aria-controls=\"documentation-nav-links\"", html);
-        Assert.Contains("class=\"documentation-nav-links\" id=\"documentation-nav-links\"", html);
-        Assert.Contains(".documentation-nav.is-enhanced:not(.is-open) .documentation-nav-links", html);
-        Assert.Contains("min-height: 44px;", html);
-        Assert.Contains("documentationNav.classList.add(\"is-enhanced\")", html);
-        Assert.Contains("setDocumentationNavOpen", html);
-    }
-
-    [Fact]
-    public async Task ProofViewer_OffersReadableMobileZoomAndPanControls()
-    {
-        var factory = sharedFactory;
-        using var client = factory.CreateClient();
-
-        var html = await client.GetStringAsync("/leaderboard");
-
-        Assert.Contains("class=\"image-viewer-stage\" tabindex=\"0\"", html);
-        Assert.Contains("class=\"image-zoom-controls\" role=\"group\"", html);
-        Assert.Contains("Zoomed for readable text. Scroll to inspect the full proof.", html);
-        Assert.Contains("const proofViewerZoomLevels = [1, 1.5, 2, 3];", html);
-        Assert.Contains("window.matchMedia('(max-width: 768px)').matches ? 2 : 0", html);
-        Assert.Contains("overflow:auto;", html);
-        Assert.Contains("touch-action:pan-x pan-y pinch-zoom;", html);
-    }
-
-    [Fact]
-    public async Task GuessMyAge_ReservesBubbleSpaceAndAllowsShortViewportRecovery()
-    {
-        var factory = sharedFactory;
-        using var client = factory.CreateClient();
-
-        var html = await client.GetStringAsync("/leaderboard");
-
-        Assert.Contains("overflow-y: auto;", html);
-        Assert.Contains("padding-top: 5.25rem; /* reserve a full row for the larger value bubble */", html);
-        Assert.Contains("#gmaBubble {", html);
-        Assert.Contains("#gmaRealBubble {", html);
-        Assert.Contains("top: 0;", html);
-        Assert.Contains("bottom: auto;", html);
-    }
-
-    [Fact]
-    public async Task HungarianChrome_LocalizesTheSharedFooterColumnHeadings()
-    {
-        var factory = sharedFactory;
-        using var client = factory.CreateClient();
-
-        var html = await client.GetStringAsync("/helstab-kihivas");
-
-        Assert.Contains(">Felfedezés<", html);
-        Assert.Contains(">Kövess minket<", html);
-        Assert.DoesNotContain(">Explore<", html);
-        Assert.DoesNotContain(">Follow<", html);
-    }
-
     [Theory]
     [InlineData("/error/502.html")]
     [InlineData("/error/503.html")]
